add product to event model
