+++ src/partner/modules/EventList/OrdersList.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent } from "react";
 import { IOrder } from "src/partner/models/Order";
 import styledComponents from "styled-components";
 import styledComponentsTS from "styled-components-ts";
-import { IProduct } from "src/partner/models/Order";
+import { IProduct } from "src/partner/models/Product";
 
 export interface IOrderTableTr {
   backgroundGray?: boolean;
